Extract number data type to file
